recursos

+++ install.sh
@@ -8,7 +8,10 @@ sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/too
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
 
-rm ~/.zshrc
+if [ -e "~/.zshrc" ]; then
+  rm ~/.zshrc
+fi
+
 ln -s ~/sophie/.zshrc ~/.zshrc
 
 echo "Se instalo zsh"
@@ -29,3 +32,5 @@ ln -s ~/sophie/.config/picom ~/.config/picom
 picom -b --config ~/.config/picom/picom.conf
 
 echo "Se instalo picom"
+
+cd resources && ./install.sh
